Extract useEditorAutoFocus hook and add clean-run to Makefile
